nba_eval/nba_game_collector_bedrock/requirements.txt
boto3>=1.34.0
google-cloud-bigquery>=3.25.0
google-auth>=2.34.0
